requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
tensorflow==2.21.0
numpy<2
pyttsx3==2.90
Werkzeug==2.3.7
Jinja2==3.1.2
mysql-connector-python>=9.0.0
